+++ packages/core/collections/users.js
@@ -4,7 +4,6 @@ Users = Meteor.users;
 if(Meteor.isServer)
 	Users._ensureIndex({"loc": "2d"});
 
-
 Users.allow({
 	update: function (userId, doc, fields, modifier) {
 		return userId && userId === doc._id;